api-gateway/requirements.txt
# api-gateway/requirements.txt
Flask>=2.0
requests>=2.25
python-dotenv>=0.19
bcrypt>=4.0.1
PyJWT>=2.8.0
pyodbc>=5.0.1